src/requirements.txt
streamlit==1.28.2
pandas==2.1.4
numpy==1.25.2
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.3.8
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
plotly==5.17.0
seaborn==0.13.0
matplotlib==3.8.2
scipy==1.11.4
scikit-learn==1.3.2
lxml==4.9.3
aixplain==0.4.0
langchain==0.0.350
tiktoken==0.5.2
chromadb==0.4.18
